Struct wee_alloc::WeeAlloc

pub struct WeeAlloc<'a> { /* fields omitted */ }

A wee allocator.

Safety

When used in unix environments, cannot move in memory. Typically not an issue if you're just using this as a static global allocator.

Methods

impl<'a> WeeAlloc<'a>

pub const INIT: Self

INIT: Self = <Self as ConstInit>::INIT

An initial const default construction of a WeeAlloc allocator.

This is usable for initializing statics that get set as the global allocator.